+++ askbot/doc/source/changelog.rst
@@ -1,6 +1,13 @@
 Changes in Askbot
 =================
 
+In development
+---------------------
+* Hide "website" and "about" section of the blocked user profiles
+  to help prevent user profile spam (Evgeny)
+* Added a function to create a custom user profile tab,
+  the feature requires access to the server (Evgeny)
+
 0.7.43 (May 14, 2012)
 ---------------------
 * User groups (Evgeny)